[CI/Build] [3/3] Reorganize entrypoints tests (#5966)
