+++ tests/entrypoints/openai/test_chat.py
@@ -14,7 +14,7 @@ import torch
 from huggingface_hub import snapshot_download
 from openai import BadRequestError
 
-from ..utils import RemoteOpenAIServer
+from ...utils import RemoteOpenAIServer
 
 # any model with a chat template should work here
 MODEL_NAME = "HuggingFaceH4/zephyr-7b-beta"
@@ -68,8 +68,6 @@ TEST_CHOICE = [
     "Python", "Java", "JavaScript", "C++", "C#", "PHP", "TypeScript", "Ruby",
     "Swift", "Kotlin"
 ]
-
-pytestmark = pytest.mark.openai
 
 
 @pytest.fixture(scope="module")
