+++ tests/entrypoints/openai/test_oot_registration.py
@@ -1,7 +1,6 @@
 import sys
 import time
 
-import pytest
 import torch
 from openai import OpenAI, OpenAIError
 
@@ -9,8 +8,6 @@ from vllm import ModelRegistry
 from vllm.model_executor.models.opt import OPTForCausalLM
 from vllm.model_executor.sampling_metadata import SamplingMetadata
 from vllm.utils import get_open_port
-
-pytestmark = pytest.mark.openai
 
 
 class MyOPTForCausalLM(OPTForCausalLM):
