+++ tests/entrypoints/openai/test_serving_chat.py
@@ -1,14 +1,10 @@
 import asyncio
 from dataclasses import dataclass
 
-import pytest
-
 from vllm.entrypoints.openai.serving_chat import OpenAIServingChat
 
 MODEL_NAME = "openai-community/gpt2"
 CHAT_TEMPLATE = "Dummy chat template for testing {}"
-
-pytestmark = pytest.mark.openai
 
 
 @dataclass
